# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ClarificationRequestResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "questions": {
            "items": {
              "type": "string"
            },
            "title": "Questions",
            "type": "array"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "requestedByUserId": {
            "format": "uuid",
            "title": "Requestedbyuserid",
            "type": "string"
          },
          "route": {
            "title": "Route",
            "type": "string"
          }
        },
        "required": [
          "id",
          "route",
          "reason",
          "questions",
          "requestedByUserId",
          "createdAt"
        ],
        "title": "ClarificationRequestResponse",
        "type": "object"
      },
...
      "CmCapabilityReviewResponse": {
        "properties": {
          "canSatisfy": {
            "title": "Cansatisfy",
            "type": "boolean"
          },
          "candidateTeams": {
            "items": {
              "$ref": "#/components/schemas/CandidateTeamResponse"
            },
            "title": "Candidateteams",
            "type": "array"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "estimatedEffort": {
            "title": "Estimatedeffort",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "managerReviewRequired": {
            "title": "Managerreviewrequired",
            "type": "boolean"
          },
          "reasoningSummary": {
            "title": "Reasoningsummary",
            "type": "string"
          },
          "requiredClarifications": {
            "items": {
              "type": "string"
            },
            "title": "Requiredclarifications",
            "type": "array"
          },
          "risks": {
            "items": {
              "type": "string"
            },
            "title": "Risks",
            "type": "array"
          },
          "suggestedCollectionRoute": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestedcollectionroute"
          },
          "suggestedCollectionSources": {
            "items": {
              "type": "string"
            },
            "title": "Suggestedcollectionsources",
            "type": "array"
          },
          "suggestedCollectionTeamId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestedcollectionteamid"
          },
          "suggestedCollectionTeamName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestedcollectionteamname"
          }
        },
        "required": [
          "id",
          "canSatisfy",
          "confidence",
          "requiredClarifications",
          "suggestedCollectionRoute",
          "suggestedCollectionTeamId",
          "suggestedCollectionTeamName",
          "suggestedCollectionSources",
          "estimatedEffort",
          "risks",
          "managerReviewRequired",
          "reasoningSummary",
          "createdAt",
          "candidateTeams"
        ],
        "title": "CmCapabilityReviewResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PriorityAssessmentResponse": {
        "properties": {
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "score",
          "tier",
          "reasons"
        ],
        "title": "PriorityAssessmentResponse",
        "type": "object"
      },
...
      "RouteApprovalRequest": {
        "properties": {
          "overrideReason": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overridereason"
          },
          "route": {
            "pattern": "^(rfa|cm)$",
            "title": "Route",
            "type": "string"
          }
        },
        "required": [
          "route"
        ],
        "title": "RouteApprovalRequest",
        "type": "object"
      },
      "RouteClarificationRequest": {
        "properties": {
          "questions": {
            "items": {
              "maxLength": 300,
              "minLength": 3,
              "type": "string"
            },
            "maxItems": 5,
            "minItems": 1,
            "title": "Questions",
            "type": "array"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          },
          "route": {
            "pattern": "^(rfa|cm)$",
            "title": "Route",
            "type": "string"
          }
        },
        "required": [
          "route",
          "reason",
          "questions"
        ],
        "title": "RouteClarificationRequest",
        "type": "object"
      },
      "RouteReasonRequest": {
        "properties": {
          "reason": {
            "maxLength": 1000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          },
          "route": {
            "pattern": "^(rfa|cm)$",
            "title": "Route",
            "type": "string"
          }
        },
        "required": [
          "route",
          "reason"
        ],
        "title": "RouteReasonRequest",
        "type": "object"
      },
...
      "RoutingQueueResponse": {
        "properties": {
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "stats": {
            "$ref": "#/components/schemas/RoutingStatsResponse"
          },
          "tickets": {
            "items": {
              "$ref": "#/components/schemas/RoutingTicketResponse"
            },
            "title": "Tickets",
            "type": "array"
          }
        },
        "required": [
          "tickets",
          "stats",
          "nextCursor"
        ],
        "title": "RoutingQueueResponse",
        "type": "object"
      },
      "RoutingStatsResponse": {
        "properties": {
          "analystAssignmentCount": {
            "title": "Analystassignmentcount",
            "type": "integer"
          },
          "clarificationCount": {
            "title": "Clarificationcount",
            "type": "integer"
          },
          "cmFallbackRate": {
            "title": "Cmfallbackrate",
            "type": "number"
          },
          "cmReviewCount": {
            "title": "Cmreviewcount",
            "type": "integer"
          },
          "rfaAcceptanceRate": {
            "title": "Rfaacceptancerate",
            "type": "number"
          },
          "rfaReviewCount": {
            "title": "Rfareviewcount",
            "type": "integer"
          },
          "routeAssessmentCount": {
            "title": "Routeassessmentcount",
            "type": "integer"
          }
        },
        "required": [
          "routeAssessmentCount",
          "rfaReviewCount",
          "cmReviewCount",
          "clarificationCount",
          "analystAssignmentCount",
          "rfaAcceptanceRate",
          "cmFallbackRate"
        ],
        "title": "RoutingStatsResponse",
        "type": "object"
      },
      "RoutingTicketResponse": {
        "properties": {
          "agentRuns": {
            "items": {
              "type": "string"
            },
            "title": "Agentruns",
            "type": "array"
          },
          "clarifications": {
            "items": {
              "$ref": "#/components/schemas/ClarificationRequestResponse"
            },
            "title": "Clarifications",
            "type": "array"
          },
          "cmReview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CmCapabilityReviewResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "managerDecisions": {
            "items": {
              "$ref": "#/components/schemas/ManagerDecisionResponse"
            },
            "title": "Managerdecisions",
            "type": "array"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "priorityAssessment": {
            "$ref": "#/components/schemas/PriorityAssessmentResponse"
          },
          "recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RouteRecommendationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "requesterUserId": {
            "format": "uuid",
            "title": "Requesteruserid",
            "type": "string"
          },
          "rfaReview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RfaCapabilityReviewResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "ticketId": {
            "format": "uuid",
            "title": "Ticketid",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "workflowPlanUpdates": {
            "items": {
              "$ref": "#/components/schemas/WorkflowPlanUpdateResponse"
            },
            "title": "Workflowplanupdates",
            "type": "array"
          }
        },
        "required": [
          "ticketId",
          "reference",
          "requesterUserId",
          "state",
          "title",
          "priority",
          "priorityAssessment",
          "rfaReview",
          "cmReview",
          "recommendation",
          "clarifications",
          "agentRuns",
          "managerDecisions",
          "workflowPlanUpdates"
        ],
        "title": "RoutingTicketResponse",
        "type": "object"
      },
...
      "SimilarRequestListResponse": {
        "properties": {
          "matches": {
            "items": {
              "$ref": "#/components/schemas/SimilarRequestResponse"
            },
            "title": "Matches",
            "type": "array"
          }
        },
        "required": [
          "matches"
        ],
        "title": "SimilarRequestListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Secure intelligence tasking and product orchestration API.",
    "title": "Istari API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/routing/cm/queue": {
      "get": {
        "operationId": "cm_queue_api_v1_routing_cm_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 25,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutingQueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cm Queue",
        "tags": [
          "routing"
        ]
      }
    },
    "/api/v1/routing/rfa/queue": {
      "get": {
        "operationId": "rfa_queue_api_v1_routing_rfa_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 25,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutingQueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rfa Queue",
        "tags": [
          "routing"
        ]
      }
    },
...
    "/api/v1/routing/{route}/release-queue": {
      "get": {
        "operationId": "release_queue_api_v1_routing__route__release_queue_get",
        "parameters": [
          {
            "in": "path",
            "name": "route",
            "required": true,
            "schema": {
              "title": "Route",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 25,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutingQueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Queue",
        "tags": [
          "routing"
        ]
      }
    },
    "/api/v1/routing/{ticket_id}": {
      "get": {
        "operationId": "routing_details_api_v1_routing__ticket_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutingTicketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Routing Details",
        "tags": [
          "routing"
        ]
      }
    },
    "/api/v1/routing/{ticket_id}/approve": {
      "post": {
        "operationId": "approve_route_api_v1_routing__ticket_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutingTicketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Route",
        "tags": [
          "routing"
        ]
      }
    },
    "/api/v1/routing/{ticket_id}/clarification": {
      "post": {
        "operationId": "request_clarification_api_v1_routing__ticket_id__clarification_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteClarificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutingTicketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Clarification",
        "tags": [
          "routing"
        ]
      }
    },
    "/api/v1/routing/{ticket_id}/reject": {
      "post": {
        "operationId": "reject_route_api_v1_routing__ticket_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteReasonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutingTicketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Route",
        "tags": [
          "routing"
        ]
      }
    },
    "/api/v1/routing/{ticket_id}/release": {
      "post": {
        "operationId": "release_product_api_v1_routing__ticket_id__release_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutingTicketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Product",
        "tags": [
          "routing"
        ]
      }
    },
    "/api/v1/routing/{ticket_id}/run": {
      "post": {
        "operationId": "run_route_reviews_api_v1_routing__ticket_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutingTicketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Route Reviews",
        "tags": [
          "routing"
        ]
      }
    },
    "/api/v1/similar-requests/routing/{ticket_id}": {
      "get": {
        "operationId": "manager_matches_api_v1_similar_requests_routing__ticket_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarRequestListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Manager Matches",
        "tags": [
          "similar requests"
        ]
      }
    },
...
  }
}